Plot bivariate predictor-response function on a new grid of points for MI BKMR

Plot bivariate predictor-response function on a new grid of points for MI BKMR

Usage

PredictorResponseBivarPair.MI(
  fit,
  y,
  Z,
  X,
  whichz1 = 1,
  whichz2 = 2,
  whichz3 = NULL,
  method = "approx",
  prob = 0.5,
  q.fixed = 0.5,
  sel = NULL,
  ngrid = 50,
  min.plot.dist = 0.5,
  center = TRUE,
  Z.MI,
  ...
)

Arguments

fit

An object containing the results returned by a the kmbayes function

y

a vector of outcome data of length n.

Z

an n-by-M matrix of predictor variables to be included in the h function. Each row represents an observation and each column represents an predictor.

X

an n-by-K matrix of covariate data where each row represents an observation and each column represents a covariate. Should not contain an intercept column.

whichz1

vector identifying the first predictor that (column of Z) should be plotted

whichz2

vector identifying the second predictor that (column of Z) should be plotted

whichz3

vector identifying the third predictor that will be set to a pre-specified fixed quantile (determined by prob)

method

method for obtaining posterior summaries at a vector of new points. Options are "approx" and "exact"; defaults to "approx", which is faster particularly for large datasets

prob

pre-specified quantile to set the third predictor (determined by whichz3); defaults to 0.5 (50th percentile)

q.fixed

vector of quantiles at which to fix the remaining predictors in Z

sel

logical expression indicating samples to keep; defaults to keeping the second half of all samples

ngrid

number of grid points to cover the range of each predictor (column in Z)

min.plot.dist

specifies a minimum distance that a new grid point needs to be from an observed data point in order to compute the prediction; points further than this will not be computed

center

flag for whether to scale the exposure-response function to have mean zero

Z.MI

Multiple Imputed Z

...

other arguments to pass on to the prediction function

Value

a data frame with value of the first predictor, the value of the second predictor, the posterior mean estimate, and the posterior standard deviation