Single variable plot for CDE.

A combination function of VarRiskSummary and riskSummary.approx for MI BKMR fits. Compare estimated h function when a single variable (or a set of variables) is at the 75th versus 25th percentile, when all of the other variables are fixed at a particular percentile

CDEVarRiskSummary.CMA(
  whichz = 1,
  BKMRfits,
  e.y = NULL,
  e.y.names = NULL,
  m.value = NULL,
  m.quant = c(0.1, 0.5, 0.75),
  m.name,
  qs.diff = c(0.25, 0.75),
  q.fixed = 0.5,
  alpha,
  sel,
  seed
)

Arguments

whichz

vector indicating which variables (columns of Z) for which the summary should be computed

BKMRfits

An object contatinint the results return by the kmbayes function

e.y

effect modifier for the outcome variable

e.y.names

column name of the effect modifier for the outcome variable

m.value

values that the mediator is set to

m.quant

values of the quantile that the mediator is set to

m.name

column name of the mediator

qs.diff

vector indicating the two quantiles q_1 and q_2 at which to compute h(z_q2)-h(z_q1)

q.fixed

a second quantile at which to compare the estimated h function

alpha

1-confidence interval

sel

a vector selecting which iterations of the fit should be retained or inference

seed

the random seed to use to evaluate the code

Value

a data frame containing the (posterior mean) estimate and posterior standard deviation of the single predictor CDE risk measures

Details

For guided examples, go to https://zc2326.github.io/causalbkmr/articles/BKMRCMA_QuickStart.html